pyproject.toml
[project]
name = "rag-based-cover-letter-cold-email-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.6",
    "faiss-cpu>=1.13.1",
    "google-api-core>=2.28.1",
    "google-genai>=1.55.0",
    "ipykernel>=7.1.0",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.3",
    "langchain-groq>=1.1.0",
    "langchain-openai>=1.1.1",
    "langgraph>=1.0.4",
    "pymupdf>=1.26.6",
    "pypdf>=6.4.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.52.1",
    "typesense>=1.3.0",
]
